Move propmts to a submodule. Refactor.
